Adds structure for an event log and basic events
- Adds style for a message/event log. The log is located to the left side the
screen on viewports larger than 750px and below the game screen on
smaller viewports.

- Adds html necessary to display the message/event log.

- Adds a ComponentController interface. Classes that implement this
will update individual parts of the UI on game events.

- Adds a MessageLogController. This implements ComponentController and
is responsible for updating the event/message log.

- Adds a components property to the GameController class, which will
store implementations of ComponentController.

- Adds a handleGameEvents() method to GameController, which routes
events to their appropriate ComponentController.

- Adds a private initializeComponents() method to GameController, which
instantiates and stores all an instance of all classes that implement
ComponentController.

- Adds logic to Character.reduceHealth() to emit a deathEvent when a
character's health falls to 0 or below.

- adds an emitEvent() method to Character class which will pass events
up to the zone the character is in.

- Adds an abstract method generateDeathMessage() to the character class.

- Adds a name prop to the NPC class and an implementation of the
generateDeathMessage() method.

- Adds an implementation of generateDeathMessage() to the Player class.

- Adds the abstract GameEvent class and the MessageEvent class.

- GameEvents have, minimally, a type, and details object which each
subclass will define. Types are a hardcoded union of strings that will
be used mainly by the GameController to decide how to handle an event,
and determine what information the event will contain.

- Alters Zone class to keep a reference to the game they are in, allowing
them to emit events to the game instance.
- - Adds emitEvent() method to Zone class.
- - Changes attack() method to emit an attack MessageEvent.

- Changes the Game class' currentZone prop to optional, to allow zones
to store a reference to the Game singleton during instantiation.

- Writes tests for all methods added to models.