Adds turn system
-- GameController Class --

- Refactors GameController to allow "space" and "numpad5" to skip turns.

- Removes zone rendering from player movement input handlers. Now
rendering  relies on the gameModel sending in events that warrant a
render or a direct call to notify();

-- Character Class --

- Adds and actionPoints and speed property. Both default to 100.

- Adds methods reduceActionPoints(), restoreAP(), and an abstract
endTurn() method. Also moves attack() method and logic from the Zone
class onto the Character class.

- Adds a getAP() method that returns the character's current AP.

-- NPC Class --

- Adds an empty endTurn() method to the NPC class to fulfill requirement
of implementing the abstract method in Character superclass. Will add
actual logic once work on AI begins.

-- Player Class --

- Changes name of move() method to handleInput(). Adds logic within to
initiate a move or an attack, as well as reduce action points on a move.

- Adds endTurn() method that calls restoreAP() and emitEvent() with an
end turn event.

-- GameEvent Class --

- Adds a GameEvent abstract class as well as the subclasses: MessageEvent,
TurnOverEvent, and MoveEvent.

- Adds static methods on the GameEvent class to allow creation of the
subclasses.

- Adds a union type of GameEventTypes to create constants representing
Event types that can be referenced across the game and used to determine
how to handle an individual event.

- Adds a message detail interface to standardize event log message
formats.

-- Zone Class --

- Removes attack() method and places it into the Character class.

- Removes logic that determined whether to move a character or begin
an attack from moveCharacter(), thus reducing the responsibility of the
method to just moving a character.

-- Game Class --

- Adds logic for handling turn over events to handleEvent().

- Adds a takeTurn() method that will notify the controller and
continue to call itself untill the player character regains enough
AP to act. Will eventually add processing for AI, status effects, etc.

-- Tests --

- Renamed MessageEvent.test.ts to GameEvent.test.ts as the focus in the
tests shifted to the abstract superclass class. The tests on the
superclass implicitly tested all salient aspects of its subclasses, so
no subclass-specific tests were written.

- Adds/refactors tests for all classes and methods mentioned above
except for GameController. Testing for that class will be implemented on
the next commit.