Adds player stats to the UI
- Adds a PlayerInfoController class implementing the ComponentController
interface that is responsible for updating DOM elements to reflect
changes in player state.

- Tracked state includes health, speed, current AP, level, name,
and equipment.

- Adds elements to the game.html file to allow for DOM manipulation

- Adds styles for the player stats UI component in \css\components\
player-info.css and imports it in the game.css file.

- Renames Controller.ts to ComponentController.ts to better reflect the
scale of the interface.

- Adds tests for the PlayerInfoController class.

- Updates relevant setters and the equipItem() method in the Player
class to call a UIChange() method which creates and emits a
PlayerUIChangeEvent to the GameController.

- Adds a levelUp() method to the Character class.

- Separates out some helper methods for the tests in the tests/public/
js/Controllers directory to assist in setting up the DOM necessary
for testing the controllers.

- Adds tests for all altered/added methods in Character and Player
classes.