Converts project to an Electron project
- Adds electron and @types/electron as dependencies to project.

- Converts project from ESModule to CommonJS for compatibility with
electron.

- Adds a "start" script that starts the project via electron and changes
project entry point to "main.js".

- ESModule syntax is retained in src and tests as TypeScript compiler
will convert the src to CommonJS and ts-node allows for tests to be run
with TS files.

- Removes all "*.js" extensions from import statements throughout
project as the conversion caused them to trigger errors during testing.
- - In other words, the extension went from being a peculiar necessity,
    to a genuine syntactical error.

- Edits tsconfig.js to output to the "/build" directory and emit
CommonJS.

- Adds a main.ts file to src/public to start up the main electron
process and renderer process.
- - This file caused TS building to emit a "js" directory that had a
    child "js" directory. This was the impetus for the above change in
    TS output directory.

- Changes the <script> tag in game.html to a single require() statement
that imports the game's starting scripts. Moves the script tag to the
bottom of the file to allow elements to load into DOM for JS targeting.

- Adds boilerplate gitignore lines from github's recommended nodejs
gitignore file.

- Adds logic to set up a basic testing scenario in the Game class'
initialize method. This method sets up some extra event handlers and
gives the player a sword and potion for manual testing purposes.