Adds stat change indicators to inventory
- Adds a +/- display when hovering over items in the inventory menu that
shows the difference between an equipped item and its potential replace-
ment.
- - IE it will show the amount of armor lost when removing a piece or
replacing it with an weaker piece, and it will show the amount gained
when equipping a stronger weapon.

- Implements the above through additional DOM elements in the stats
table of the InventoryController class as well as event listeners added
onto the rows of the inventory table and the rows with items of the
equipped table.

- Adds a .5em left margin to the element responsible for displaying
the difference to inventory.css.

- Adds getBaseDamage method to the Character class to get the raw damage
needed to calculate the difference between an equipped weapon and a
potential replacement.